perf: paginate user directory by 15
